modified simulation script and added gnuplot

+++ simul.sh
@@ -8,14 +8,22 @@ ELEMENTS=${#args[@]}
  
 data=$1
 ber=$2
+step=$3
 
+
+if [ $# -eq 0 ]; then
+    echo "usage : ./simul.sh <taille données> <BER> <step>"
+    exit
+fi
+
+echo "Starting simulation..."
 rm -f results.txt
 
 # for each data paylaod
-for (( i=2;i<$ELEMENTS;i++)); do
+for (( i=200;i<=1476;i+=$step)); do
 	cd src/
-	results=$(python main.py -q -P ${args[${i}]} -r $data $ber)
-	read results< <(echo "$results" | tail -n1)
+	results=$(python3 main.py -q -P $i -r $data $ber)
+ 	read results< <(echo "$results" | tail -n1)
 	echo $results
 	cd ../
 	touch results.txt
@@ -23,8 +31,7 @@ for (( i=2;i<$ELEMENTS;i++)); do
 	echo $results$'\r' >> results.txt
 done
 
+gnuplot -p -e "plot 'results.txt' u 3:5 w l"
+
 echo "done!"
-cmd /k
-
-
